+++ docs/conf.py
@@ -35,7 +35,7 @@ author = 'Jochen Gerhäußer'
 # The short X.Y version.
 version = '1.1'
 # The full version, including alpha/beta/rc tags.
-release = '1.1a0'
+release = '1.1'
 
 # -- General configuration ---------------------------------------------------